refactor(frontend): use service factories
