+++ frontend/server/api/blog/test.ts
@@ -1,9 +1,11 @@
-import { blogService } from '~/services/blog.services'
+import { useBlogService } from '~/services/blog.services'
 
 /**
  * Test endpoint to debug blog data
  */
 export default defineEventHandler(async _event => {
+  const blogService = useBlogService()
+
   try {
     // Get raw data from the external API
     const response = await blogService.getArticles()
