+++ frontend/services/content.services.ts
@@ -4,23 +4,18 @@ import type { XwikiContentBlocDto } from '~/src/api'
 /**
  * Content service for fetching HTML blocs from the backend
  */
-export class ContentService {
-  private readonly api: ContentApi
-
-  constructor() {
-    const config = useRuntimeConfig()
-    const apiConfig = new Configuration({ basePath: config.apiUrl })
-    this.api = new ContentApi(apiConfig)
-  }
+export const useContentService = () => {
+  const config = useRuntimeConfig()
+  const apiConfig = new Configuration({ basePath: config.apiUrl })
+  const api = new ContentApi(apiConfig)
 
   /**
    * Retrieve a content bloc by its identifier
    * @param blocId - XWiki bloc identifier
    */
-  async getBloc(blocId: string): Promise<XwikiContentBlocDto> {
-    return await this.api.contentBloc({ blocId })
+  const getBloc = async (blocId: string): Promise<XwikiContentBlocDto> => {
+    return await api.contentBloc({ blocId })
   }
-}
 
-// Export singleton instance
-export const contentService = new ContentService()
+  return { getBloc }
+}
