+++ frontend/services/index.ts
@@ -1,4 +1,4 @@
-// Export all services
-export { blogService } from '~/services/blog.services'
-export { contentService } from '~/services/content.services'
+// Export service factories and singletons
+export { useBlogService } from '~/services/blog.services'
+export { useContentService } from '~/services/content.services'
 export { authService } from '~/services/auth.services'
